Options class for configuration variables
